Rename command folders
