+++ ledger_bot/LedgerBot.py
@@ -7,7 +7,7 @@ import discord
 from apscheduler.schedulers.asyncio import AsyncIOScheduler
 
 from .clients import ExtendedClient, ReactionRolesClient, TransactionsClient
-from .process_dm import is_dm, process_dm
+from .commands_dm import is_dm, process_dm
 from .process_message import process_message
 from .reminder_manager import ReminderManager
 from .services import Service
