+++ ledger_bot/commands_dm/__init__.py
@@ -1,0 +1,6 @@
+"""Functions to process DM messages."""
+
+from .is_dm import is_dm
+from .process_dm import process_dm
+
+__all__ = ["process_dm", "is_dm"]
