+++ ledger_bot/commands_slash/__init__.py
@@ -1,0 +1,5 @@
+"""Module containing all our 'slash commands'."""
+
+from .setup_slash import setup_slash
+
+__all__ = ["setup_slash"]
