+++ ledger_bot/commands_slash/command_new_sale.py
