+++ ledger_bot/commands_slash/command_ping.py
