--- ledger_bot/process_dm/__init__.py
@@ -1,10 +1,0 @@
-"""Functions to process DM messages."""
-
-from .command_help import command_help
-from .command_list import command_list
-from .command_version import command_version
-from .get_dm_channel import get_dm_channel
-from .is_dm import is_dm
-from .process_dm import process_dm
-
-__all__ = ["process_dm", "is_dm"]
