--- ledger_bot/slash_commands/__init__.py
@@ -1,10 +1,0 @@
-"""Module containing all our 'slash commands'."""
-
-from .command_add_user import command_add_user
-from .command_hello import command_hello
-from .command_help import command_help
-from .command_list import command_list
-from .command_new_sale import command_new_sale
-from .setup_slash import setup_slash
-
-__all__ = ["setup_slash"]
